requirements.txt
arrow==1.2.3
asgiref==3.6.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-filter==23.2
django-geojson==4.0.0
django-gisserver==1.2.5
django-leaflet==0.29.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
Flask==2.3.2
GDAL @ file:///C:/Users/cfick/dev/mymap/mymap_project/GDAL-3.4.3-cp310-cp310-win_amd64.whl#sha256=9893da5532a6c5dadc1fc3bc402ae29d161b7cf190d778f129da84d0e50baa29
geos==0.2.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lru-dict==1.1.8
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.3
orjson==3.8.12
Pillow==9.5.0
proj==0.2.0
psycopg2==2.9.6
pycparser==2.21
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.14.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-toolbelt==1.0.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.3
wrapt==1.15.0
